+++ gemp-lotr/gemp-lotr-server/src/main/java/com/gempukku/lotro/bots/simulation/GameResult.java
@@ -1,4 +1,4 @@
-package com.gempukku.lotro.simulation;
+package com.gempukku.lotro.bots.simulation;
 
 public enum GameResult {
     P1_WON, P2_WON
