+++ gemp-lotr/gemp-lotr-server/src/main/java/com/gempukku/lotro/bots/simulation/SimpleBatchSimulationRunner.java
@@ -1,4 +1,6 @@
-package com.gempukku.lotro.simulation;
+package com.gempukku.lotro.bots.simulation;
+
+import com.gempukku.lotro.bots.BotPlayer;
 
 public class SimpleBatchSimulationRunner implements SimulationRunner{
     private final Simulation simulation;
