+++ gemp-lotr/gemp-lotr-server/src/main/java/com/gempukku/lotro/bots/simulation/Simulation.java
@@ -1,4 +1,6 @@
-package com.gempukku.lotro.simulation;
+package com.gempukku.lotro.bots.simulation;
+
+import com.gempukku.lotro.bots.BotPlayer;
 
 public interface Simulation {
     GameResult simulateGame(BotPlayer bot1, BotPlayer bot2);
