+++ gemp-lotr/gemp-lotr-server/src/main/java/com/gempukku/lotro/game/LotroServer.java
@@ -9,8 +9,8 @@ import com.gempukku.lotro.db.DeckDAO;
 import com.gempukku.lotro.hall.GameSettings;
 import com.gempukku.lotro.logic.timing.GameResultListener;
 import com.gempukku.lotro.logic.vo.LotroDeck;
-import com.gempukku.lotro.simulation.BotPlayer;
-import com.gempukku.lotro.simulation.BotService;
+import com.gempukku.lotro.bots.BotPlayer;
+import com.gempukku.lotro.bots.BotService;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import java.util.*;
